LinkedIn_summarizer/requirements.txt
streamlit
google_generativeai
python-dotenv
vertexai